+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/sales/model/dto/SalesDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class SalesDTO {
     private Integer salesId;
     private Integer storeId;
-    private Integer userId;
+    private Long userId;
     private String userName;
     private String userPhone;
     private LocalDate salesDate;
@@ -27,7 +27,7 @@ public class SalesDTO {
     public SalesDTO() {
     }
 
-    public SalesDTO(Integer salesId, Integer storeId, Integer userId, String userName, String userPhone, LocalDate salesDate, LocalDate shippingDueDate, SalesStatus salesStatus, LocalDateTime salesCreatedAt, LocalDateTime salesUpdatedAt) {
+    public SalesDTO(Integer salesId, Integer storeId, Long userId, String userName, String userPhone, LocalDate salesDate, LocalDate shippingDueDate, SalesStatus salesStatus, LocalDateTime salesCreatedAt, LocalDateTime salesUpdatedAt) {
         this.salesId = salesId;
         this.storeId = storeId;
         this.userId = userId;
@@ -56,11 +56,11 @@ public class SalesDTO {
         this.storeId = storeId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
